Remove resource smoke test to functional test
Also added an equivalent test for clients.
